Allow the user to specify the cancellation event.
In version 1.0 of the application, a screen unblank event would
cancel the recurring notification.  With the current code base,
that is now switched to using the ACTION_USER_PRESENT intent, which
is broadcast when the screen is unlocked.  There are two issues with
this behaviour:
1) it is a departure from what users have today
2) if there is no lock screen configured, no ACION_USER_PRESENT intent
   will be broadcast.
This patch makes the behaviour configurable, as there is no way to
discover how the user has set his/her phone up.

Note, there is still a need for a follow-on patch to allow the user
to specify the lock screen timeout so that we don't continue to send
notification when the user is actually present.